requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.9.0
python-dotenv==1.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0
PyJWT==2.9.0
python-dateutil==2.9.0
httpx==0.27.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
cryptography==43.0.0
